Move the Language.jsx to a separate folder 'languages'.
