+++ src/components/form/components/languages/ui/Language.jsx
@@ -1,6 +1,6 @@
 import React, { useContext } from "react";
-import {ResumeContext} from "../../builder";
-import FormButton from "./FormButton";
+import {ResumeContext} from "../../../../builder";
+import FormButton from "./../../FormButton";
 
 const Language = () => {
   const { resumeData, setResumeData } = useContext(ResumeContext);
